Move e2e tests from protractorE2E to e2e
