+++ scenarioo-client/gulpfile.js
@@ -137,7 +137,7 @@ gulp.task('test-watch', ['environmentConstants'], function (done) {
  * Run protractor ui tests (without generating Scenarioo documentation)
  */
 gulp.task('test-e2e', function () {
-    gulp.src(['./test/protractorE2E/specs/**/*.js'])
+    gulp.src(['./test/e2e/specs/**/*.js'])
         .pipe(protractor({
             configFile: './protractor-e2e.conf.js'
         }))
@@ -152,7 +152,7 @@ gulp.task('webdriver_update', webdriver_update); // eslint-disable-line no-undef
  * Run protractor ui tests and generate Scenarioo documentation from them.
  */
 gulp.task('test-e2e-scenarioo', function () {
-    gulp.src(['./test/protractorE2E/specs/**/*.js'])
+    gulp.src(['./test/e2e/specs/**/*.js'])
         .pipe(protractor({
             configFile: './protractor-e2e-scenarioo.conf.js'
         }))
