+++ scenarioo-client/prepareProtractor.js
@@ -6,7 +6,7 @@ var jasmineReporters = require('jasmine-reporters');
 function prepareProtractorForE2ETests() {
 
     require("ts-node").register({
-        project: "test/protractorE2E/tsconfig.json"
+        project: "test/e2e/tsconfig.json"
     });
 
     setupJasmineXmlReporters();
@@ -32,7 +32,7 @@ function prepareProtractorForE2ETests() {
         // this is also needed if scenarioo documenztation is not enabled
         var scenarioo = require('scenarioo-js');
         scenarioo.setupFluentDsl();
-        require('./test/protractorE2E/labelDefinitions');
+        require('./test/e2e/labelDefinitions');
     }
 
     function disableAnimations() {
