+++ scenarioo-client/test/e2e/specs/application_startup.ts
