+++ scenarioo-client/test/e2e/specs/browse_object_details.ts
