+++ scenarioo-client/test/e2e/specs/browse_page_variants.ts
