+++ scenarioo-client/test/e2e/specs/configure_label_colors.ts
