+++ scenarioo-client/test/e2e/specs/create_sketch.ts
