+++ scenarioo-client/test/e2e/specs/full_text_search.ts
