+++ scenarioo-client/test/e2e/specs/step_screen_annotations.ts
