+++ scenarioo-client/test/e2e/webPages/labelConfigurationsPage.ts
