+++ scenarioo-client/test/e2e/webPages/objectDetailsPage.ts
